Change DBM tool from mysql to postgresql
This change occurred because of an error I kept on having while using mysql.
The error was the following:
BLOB/TEXT column 'product_name' used in key specification without a key length.
After some investigation it appeared to be an bug that can not be fixed so it was suggested to move to an alternative DBM tool (https://code.djangoproject.com/ticket/2495).